% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_to_zero.R
\name{fix_to_zero}
\alias{fix_to_zero}
\title{Fix a Free Parameter To Zero And Fit a Model Again}
\usage{
fix_to_zero(fit, par_id, store_fit = FALSE)
}
\arguments{
\item{fit}{A \code{lavaan}-class object.}

\item{par_id}{An integer. The row
number of the free parameter in the
parameter table of \code{fit} to be
fixed.}

\item{store_fit}{Logical. If \code{TRUE},
\code{fit} will be stored in the output.
Default is \code{FALSE}.}
}
\value{
A \code{fix_to_zero}-class object, which
is a list with these elements:call
\itemize{
\item \code{fit0} is the \code{lavaan} output of the
refitted object. \code{NA} if the fit
failed for some reasons. A model
must converged, passed \code{lavaan}'s
post check, variance covariance
matrix of estimates can be computed,
and has one more degree of freedom
than the original model to be
considered acceptable.
\item \code{fit1} is the original \code{lavaan}
output if \code{store_fit} is \code{TRUE}. It
is \code{NULL} otherwise
\item \code{par_id} is the row number of the
designated free parameter.
\item \code{call} is the original call to this
function.
\item \code{ptable0} is the parameter table
of with the designated parameter fixed
to zero. It can be used for diagnostic
purpose if the fit failed.
\item \code{fit0_error} is the error
message in refitting, if any. If
no error, it is `NA``.
\item \code{vcov_ok} is \code{TRUE} if the variance
covariance matrix of the estimates
can be computed without error nor
warning. \code{FALSE} otherwise.
\item \code{vcov_msg} is the message generated
when using \code{\link[lavaan:lavInspect]{lavaan::lavInspect()}} to
get the variance-covariance matrix
of the parameter estimates of the
refitted model. If \code{TRUE}, then no
error nor warning. Can be used for
diagnostic purposes.
\item \code{converged}: Whether the modified
model converged.
\item \code{post_check_passed}: Whether the
solution of the modified model passed
\code{lavaan}s post check.
\item \code{fit_not_ok}: If the fit failed
for some reasons, the fit object,
if available, is stored in this
element rather than in \code{fit0}. such
that the fit object can be retrieved
for diagnostic purposed if necessary.
\item \code{df_diff_one}: Whether the
difference in model degrees of
freedom between the modified model
and the original model is one.
}
}
\description{
It fixes a designated free parameter
in a lavaan object to zero and refit the model.
}
\details{
It modifies the parameter
table of a \code{lavaan}-class object and
then fits the model again.
}
\examples{
library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"
fit <- sem(mod, data_sem16)
# Fix the factor covariance to zero
out <- fix_to_zero(fit, par_id = 15)
summary(out$fit0)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
