% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{Fix a Free Parameter to Zero and do LR Test}
\usage{
lrt(fit, par_id)
}
\arguments{
\item{fit}{A \code{lavaan}-class object.}

\item{par_id}{An integer. The row
number of the free parameter in the
parameter table of \code{fit} to be
fixed.}
}
\value{
A \code{lrt}-class object, which is a
list with the following elements:
\itemize{
\item \code{lrt}: The output of \code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}}.
\item \code{par_id}: The row number of the
designated free parameters.
\item \code{fit0}: The \code{lavaan} output of the
modified model, with the designated
free parameter fixed to zero.
\item \code{fit1}: The original \code{lavaan}
output.
\item \code{call}: The call to this function.
}
}
\description{
Fix the designated
free parameter to zero and do a
likelihood ratio test.
}
\details{
It fixes the designated
free parameter in a \code{lavaan} output,
refit the model, and do a likelihood
ratio test comparing this model with
the original model.
}
\examples{
library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"
fit <- sem(mod, data_sem16)
# Fix the factor covariance to zero
out <- lrt(fit, par_id = 15)
out$lrt
parameterEstimates(fit)[15, ]
parameterEstimates(out$fit0)[15, ]

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
