% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlrtp-package.R
\docType{package}
\name{semlrtp-package}
\alias{semlrtp}
\alias{semlrtp-package}
\title{semlrtp: Likelihood Ratio Test P-Values for Structural Equation Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Computes likelihood ratio test (LRT) p-values for free parameters in a structural equation model. Currently supports models fitted by the 'lavaan' package by Rosseel (2012) \doi{10.18637/jss.v048.i02}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semlrtp/}
  \item Report bugs at \url{https://github.com/sfcheung/semlrtp/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com} (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
}

}
\keyword{internal}
