% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_to_zero.R
\name{fix_to_zero}
\alias{fix_to_zero}
\title{Fix a Free Parameter To Zero And Fit a Model Again}
\usage{
fix_to_zero(fit, par_id, store_fit = FALSE)
}
\arguments{
\item{fit}{A \code{lavaan}-class object.}

\item{par_id}{An integer. The row
number of the free parameter in the
parameter table of \code{fit} to be
fixed.}

\item{store_fit}{Logical. If \code{TRUE},
\code{fit} will be stored in the output.
Default is \code{FALSE}.}
}
\value{
A \code{fix_to_zero}-class object, which
is a list with the element \code{fit0}
storing the \code{lavaan} output of the
refitted object, \code{fit1} storing the
original \code{lavaan} output if requested
(\code{NULL} otherwise), \code{par_id} storing
the row number of the designated free
parameter, and \code{call} storing the
original call.
}
\description{
It fixes a designated free parameter
in a lavaan object to zero and refit the model.
}
\details{
It modifies the parameter
table of a \code{lavaan}-class object and
then fits the model again.
}
\examples{
library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"
fit <- sem(mod, data_sem16)
# Fix the factor covariance to zero
out <- fix_to_zero(fit, par_id = 15)
summary(out$fit0)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
