% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{Fix a Free Parameter to Zero and do LR Test}
\usage{
lrt(fit, par_id, store_fit = FALSE, group = NULL)
}
\arguments{
\item{fit}{A \code{lavaan}-class object.}

\item{par_id}{It can be an integer.
or a string. If it is an integer,
it should be the row
number of the free parameter in the
parameter table of \code{fit} to be
fixed. If it is a string, it must be
a valid \code{lavaan} model syntax, or
the label of a labelled parameter.}

\item{store_fit}{Logical. If \code{TRUE},
\code{fit} will be stored in the output.
Default is \code{FALSE}.}

\item{group}{If a model syntax
is used in \code{par_id} and the model
is a multigroup model, this should
be either the group label or the
group number of the parameter.}
}
\value{
A \code{lrt}-class object, which is a
list with the following elements:
\itemize{
\item \code{lrt}: The output of \code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}}.
If there is an error message or
warning, it is set to \code{NA}.
\item \code{par_id}: The row number of the
designated free parameters.
\item \code{fit1}: The original \code{lavaan}
output, if \code{store_fit} is \code{TRUE}.
\code{NA} otherwise.
\item \code{fix_to_zero}: The output of
\code{fit_to_zero()}.
\item \code{call}: The call to this function.
\item \code{lrt_status}: Integer. If 0, then
there is no error nor warning
in the likelihood ratio test and
\code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}} returns a
table (\code{data.frame}) of the test.
If -1, then something is wrong,
e.g., an error or warning occurred.
\item \code{lrt_msg}: If something is wrong
when doing the likelihood ratio
test, this is the error or warning
message. If no error nor warning,
this is \code{NA}.
}
}
\description{
Fix the designated
free parameter to zero and do a
likelihood ratio test.
}
\details{
It fixes the designated
free parameter in a \code{lavaan} output,
refit the model, and do a likelihood
ratio test comparing this model with
the original model.
}
\examples{
library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"
fit <- sem(mod, data_sem16)
# Fix the factor covariance to zero
out <- lrt(fit, par_id = 15)
out$lrt
parameterEstimates(fit)[15, ]
parameterEstimates(out$fix_to_zero$fit0)[15, ]

# Can use model syntax for par_id

out <- lrt(fit, par_id = "f1 =~ x3")
out$lrt

}
\seealso{
\code{\link[=print.lrt]{print.lrt()}}
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
